Luu quang duc

+++ database/migrations/2025_04_11_184922_add_order_code_don_hang_table.php
@@ -11,9 +11,11 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('don_hang', function (Blueprint $table) {
-            $table->text('orderCode')->nullable()->after('id');
-        });
+        if (!Schema::hasColumn('don_hang', 'orderCode')) { // Kiểm tra nếu cột chưa tồn tại
+            Schema::table('don_hang', function (Blueprint $table) {
+                $table->text('orderCode')->nullable()->after('id');
+            });
+        }
     }
 
     /**
@@ -21,8 +23,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('don_hang', function (Blueprint $table) {
-            $table->dropColumn('orderCode');
-        });
+        if (Schema::hasColumn('don_hang', 'orderCode')) { // Kiểm tra nếu cột tồn tại
+            Schema::table('don_hang', function (Blueprint $table) {
+                $table->dropColumn('orderCode');
+            });
+        }
     }
 };
